requirements.txt
transformers==4.35.2
langchain==0.1.13
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
regex>=2023.10.0
python-dotenv>=1.0.0
rich>=13.3.0
colorama>=0.4.6
faiss-cpu>=1.8.1
sentence-transformers>=2.5.0
huggingface_hub>=0.17.4
openai>=0.27.8
langchain-openai>=0.0.7
langchain-huggingface>=0.0.7
langchain-community>=0.0.7
langchain_core>=0.1.13
pydantic>=2.10.7
fastapi>=0.100.0
uvicorn>=0.23.2
requests>=2.31.0
# Optional: If you plan to use Docker, include the following line
# docker>=6.1.3

# To set up and run the project, follow these steps:
#python -m venv venv
#.\venv\Scripts\activate
#pip install --upgrade pip
#pip install -r requirements.txt

#python AgentRoot.py

#pip install langchain langchain-community transformers huggingface_hub openai python-dotenv
#pip install torch
#pip install torch --index-url https://download.pytorch.org/whl/cpu
#pip install scikit-learn pandas regex
#For RAG tool
#pip install faiss-cpu
# pip install sentence-transformers
#pip install -U langchain-openai langchain-huggingface langchain-community
#pip install colorama
#pip install rich

